Fix PR comments (part 2)
