+++ app/src/main/java/com/example/ermolaenkoalex/nytimes/ui/newsedit/NewsEditActivity.java
@@ -117,16 +117,13 @@ public class NewsEditActivity extends BaseActivity implements NewsEditView {
     }
 
     @Override
-    public void close(@IdRes int errorMessage) {
-        showErrorMassage(errorMessage);
+    public void close() {
         finish();
     }
 
     @Override
-    public void showErrorMassage(@IdRes int errorMessage) {
-        if (errorMessage != 0) {
-            Toast.makeText(this, errorMessage, Toast.LENGTH_LONG).show();
-        }
+    public void showErrorMessage(@IdRes int errorMessage) {
+        Toast.makeText(this, errorMessage, Toast.LENGTH_LONG).show();
     }
 
     @Override
